--- app/Resources/views/services/articles.html.twig
@@ -1,1 +1,0 @@
-{% extends 'base.html.twig' %}
