+++ src/AppBundle/Controller/HomeController.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace AppBundle\Controller;
+
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+
+/**
+ * Class HomeController
+ * @package AppBundle\Controller
+ */
+class HomeController extends Controller
+{
+    /**
+     * @Route("/", name="homepage")
+     * @return \Symfony\Component\HttpFoundation\Response
+     */
+    public function homeAction()
+    {
+        $em = $this->getDoctrine()->getManager();
+        $planets = $em->getRepository('AppBundle:Planet')->findAll();
+        $news = $em->getRepository('AppBundle:Article')->findAll();
+
+        return $this->render('home/homepage.html.twig', [
+            'planets' => $planets,
+            'news' => $news
+        ]);
+    }
+}
