+++ src/AppBundle/Tests/Controller/HomeControllerTest.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace AppBundle\Tests\Controller;
+
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+
+class HomeControllerTest extends WebTestCase
+{
+}
