+++ src/AppBundle/Tests/Controller/NewsControllerTest.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace AppBundle\Tests\Controller;
+
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+
+class NewsControllerControllerTest extends WebTestCase
+{
+}
